safari-600.7-branch/LayoutTests/storage/indexeddb/mozilla/create-objectstore-null-name-expected.txt

Test IndexedDB's creating object store with null name

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


indexedDB = self.indexedDB || self.webkitIndexedDB || self.mozIndexedDB || self.msIndexedDB || self.OIndexedDB;

dbname = "create-objectstore-null-name.html"
indexedDB.deleteDatabase(dbname)
indexedDB.open(dbname)
objectStore = db.createObjectStore(null);
PASS objectStore.name is 'null'
PASS successfullyParsed is true

TEST COMPLETE
